// PP-CoE-CodeApp/src/generated/models/PowerPlatformforAdminsV2Model.ts
export interface MgGovEnvironmentValues {
  // The ID of the environment.
  id?: string;
  // The type of the environment.
  type?: string;
  // The name of the environment.
  name?: string;
}

export interface MgGovErrorResponse {
  error?: {
    // Error code.
    code?: string;
    // Error message.
    message?: string;
  };
}

export type MgGovFilterType = string;

export interface MgGovODataResponse {
  // The array of Rule Sets.
  value?: RuleSetDto[];
  // The link to the next page of results.
  "@odata.nextLink"?: string;
}

export interface MgGovPolicyEnvironmentFilter {
  type?: MgGovFilterType;
  // The environment information.
  values?: MgGovEnvironmentValues[];
}

export type MgGovResourceType = string;

export interface MgGovRule {
  // The rule ID.
  id?: string;
  // The rule value.
  value?: string;
}

export type MgGovRuleSetType = string;
// ...
export interface Policy {
  // The unique identifier of the policy assignment.
  id?: string;
  // The unique identifier of the tenant.
  tenantId?: string;
  // The name of the policy.
  name?: string;
  // The date and time when the policy was last modified.
  lastModified?: string;
  ruleSets?: RuleSet[];
  // The number of rule sets associated with this policy.
  ruleSetCount?: number;
}
// ...
export interface RuleSet {
  // The unique identifier of the rule set.
  id?: string;
  // The version of the rule set.
  version?: string;
  // The inputs for the rule set, which may vary based on the rule.
  inputs?: Record<string, unknown>;
}

export interface RuleSetDto {
  // The ID of the Rule Set.
  id?: string;
  // The last modified timestamp.
  lastModified?: string;
  environmentFilter?: MgGovPolicyEnvironmentFilter;
  // The Rule Set parameters.
  parameters?: RuleSetParameters[];
}

export interface RuleSetParameters {
  type: MgGovRuleSetType;
  resourceType: MgGovResourceType;
  // The rule values.
  value?: MgGovRule[];
}